+++ CHANGELOG.md
@@ -7,6 +7,7 @@ is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 - ESLint
+- Playwright testing
 
 
 ## 1.1.2 - 2026-03-09